Updates and clean up (#144)
* Refactoring review

Bring in changes from phunkyfish review of 6.2.0

* Read should be ssize_t

Read length is size_t so return should not be int

* Client timeshift rework

Timeout is now based on time a user can control once data is available.

+++ src/buffers/ClientTimeshift.h
@@ -60,7 +60,7 @@ namespace timeshift {
     {
       return m_inputHandle.GetPosition();
     }
-    virtual int Read(byte *buffer, size_t length) override
+    virtual ssize_t Read(byte *buffer, size_t length) override
     {
       ssize_t dataLen = m_inputHandle.Read(buffer, length);
       if (m_complete && dataLen == 0)
